fix: apply four architecture corrections to ClaimVision
Architecture corrections applied:
1. Production-minded terminology: replaced enterprise-grade with
   production-minded across README, SKILL.md, EXECUTION_ARCHITECTURE.md

2. Pure PostgreSQL database manager: removed embedded_postgres fallback
   from database/db_manager.py. DatabaseManager now connects exclusively
   via DATABASE_URL or individual DB_* env vars.
   Canonical environment: PostgreSQL 16 via Docker Compose.

3. Provider-level group split: replaced sklearn train_test_split (row-level)
   with provider_group_split() in src/model_pipeline.py.
   No provider appears in both train and test sets.
   Assertion: set(train_ids).isdisjoint(set(test_ids)) - PASSED
   Split: 4329 train providers / 1081 test providers.

4. Dynamic scale_pos_weight: derived from training providers only.
   Neg=3924 / Pos=405 = 9.6889. Never from test set.

Updated metrics (provider-level group split):
  ROC-AUC: 0.9557  PR-AUC: 0.7369
  Tuned F1: 0.6567  Precision: 66.00%  Recall: 65.35%

Added: docker-compose.yml PostgreSQL 16-alpine
Tests: 13 passed, 5 skipped (DB tests skip gracefully without PostgreSQL)

+++ database/queries/fraud_patterns.sql
@@ -1,5 +1,5 @@
 -- =====================================================================
--- Section 4: Suspicious Fraud Patterns & Behavioral Profiling
+-- Section 4: Suspicious Fraud Patterns & Behavioral Profiling (PostgreSQL)
 -- =====================================================================
 
 -- Q14: Phantom Billing: Same Beneficiary Billed Multiple Times on Same Date
@@ -8,7 +8,7 @@ SELECT
     bene_id,
     claim_start_date,
     COUNT(claim_id) AS same_day_claim_count,
-    ROUND(SUM(reimbursed_amount), 2) AS total_same_day_payout
+    ROUND(SUM(reimbursed_amount)::numeric, 2) AS total_same_day_payout
 FROM fact_claims_unified
 GROUP BY provider_id, bene_id, claim_start_date
 HAVING COUNT(claim_id) > 1
@@ -22,7 +22,7 @@ SELECT
     p.fraud_label,
     c.length_of_stay,
     c.reimbursed_amount,
-    ROUND(c.reimbursed_amount / NULLIF(c.length_of_stay, 0), 2) AS reimbursement_per_day,
+    ROUND((c.reimbursed_amount / NULLIF(c.length_of_stay, 0))::numeric, 2) AS reimbursement_per_day,
     c.admit_diagnosis_code
 FROM fact_inpatient_claims c
 JOIN dim_providers p ON c.provider_id = p.provider_id
@@ -37,7 +37,7 @@ SELECT
     p.fraud_label,
     COUNT(DISTINCT c.claim_id) AS total_claims,
     COUNT(DISTINCT c.attending_physician) AS distinct_physicians,
-    ROUND(CAST(COUNT(DISTINCT c.attending_physician) AS FLOAT) / COUNT(DISTINCT c.claim_id), 3) AS physician_to_claim_ratio
+    ROUND(CAST(COUNT(DISTINCT c.attending_physician) AS NUMERIC) / COUNT(DISTINCT c.claim_id), 3) AS physician_to_claim_ratio
 FROM fact_claims_unified c
 JOIN dim_providers p ON c.provider_id = p.provider_id
 GROUP BY c.provider_id, p.fraud_label
@@ -49,8 +49,8 @@ LIMIT 15;
 SELECT
     primary_diagnosis_code,
     COUNT(*) AS diagnosis_occurrences,
-    ROUND(SUM(reimbursed_amount), 2) AS total_amount,
-    ROUND(AVG(reimbursed_amount), 2) AS avg_claim_cost
+    ROUND(SUM(reimbursed_amount)::numeric, 2) AS total_amount,
+    ROUND(AVG(reimbursed_amount)::numeric, 2) AS avg_claim_cost
 FROM fact_claims_unified
 WHERE primary_diagnosis_code IS NOT NULL
 GROUP BY primary_diagnosis_code
